Added Pages: Home About Gallery Contact Us
Added a footer with social media links as well as a 'back to top' button.

Home:
Scrolling images and general information about the society.

About:
Essentially a 'meet the committee' page. Need to add some more info about the society, its history etc.

Contact us:
Included an email (physical@csc.tcd.ie) as well as mailing address.

Gallery:
A page with a Gallery of images from events and brief descriptions of the events. Easily editable.

The Library page is still under construction, and the Join is just a link to the trinity society signups page.

Removed herobrine etc